requirements.txt
Django==1.4.2
MySQL-python==1.2.4c1
distribute==0.6.28
djangorestframework==2.1.6
wsgiref==0.1.2
